refactor(formatter): share close keyword scanning
